Modified Review model and rating to integer

+++ app/src/main/java/hr/fvlahov/shows_franko_vlahov/model/Review.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 data class Review(
     val id: String,
-    val rating: Float,
+    val rating: Int,
     val review: String,
     val reviewerName: String,
     @DrawableRes val reviewerImageResourceId: Int
